Fix up restore code
It was quite broken. The indentation was wrong (and thus everything was
being called the wrong number of times)

Use dict.get() only for things that we can handle being absent, lists
and dicts. For other cases, we want to catch a KeyError.

Pass the storageobjects in as a list not a generator, so they can be
iterated multiple times.

Save tpg_lun for Mapped luns. This lets us maintain mappings properly
on restore. Also require index val for tpg luns -- not doing so would lead
to randomly assigned indexes, which then could result in mapped luns
being assigned to the wrong tpg lun!

Signed-off-by: Andy Grover <agrover@redhat.com>