add has_sessions to FabricModule, Target and TPG, close #10
Implemented as a property, backed by _check_for_sessions(),
since no parameters are necessary.

At the base level (NodeACL) it actually creates a Session object and
fills it with information from configfs.
This object is then thrown away.

This stems from the fact, that you have to parse the info part of the
acl even if you only want to check if there is a session.

In contrast to returning a list of open sessions and checking if it has
a length > 0, _check_for_esssions() will stop if it finds any session in
the current configfs folder.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>