WWNs need not exist to be valid
One issue with FCoE (and perhaps other fabrics) is that the target
configuration code may be run before the underlying interface is up.
This should be OK, since the tcm_fc code monitors interfaces as they are
added and will tie them with the target instance. But, rtslib has been
preventing this, and it sounds like it shouldn't be.

Reported-by: Mark Rustad <mark.d.rustad@intel.com>
Signed-off-by: Andy Grover <agrover@redhat.com>