rtslib: don't try to be smart about module loading
Any driver in the Linux kernel can be built in, so don't try to be smart
to check for possible modules.  Just try to load a driver and ignore the
return value - either it was built in or loaded and things are fine,
or it's not available at all and the code needs to cope with this at
a higher level.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Andy Grover <agrover@redhat.com>