Use actual device name for block object size
LVM lets you rename LVs while they're in use. This results in bso.size
returning None if the backing LV is renamed, or the wrong size if another
LV has been renamed to the old LV's name, since we were using udev_path.

Use device name, which does not change as long as we have the backstore
open.

Reported-by: Xiaowei Li <xiaoli@redhat.com>
Signed-off-by: Andy Grover <agrover@redhat.com>