Add a cache for backstore lookups
For each new SO we need to verify there isn't an existing one with the
same (type, name). Unfortunately these are nested in directories, so
adding each new SO requires more directory searches. AFAIK there's no
way around this without changing configfs layout.

This change adds a backstore cache. We still resort to glob() for new
SOs (it's still O(n^2)) but avoiding this for SOs we've seen before really
seems to help a lot.

Signed-off-by: Andy Grover <agrover@redhat.com>