Preload entire backstore cache on bs_cache miss
Instead of globbing for names one-at-a-time, load the entire series of
dirp/name->index mappings into bs_cache on a cache miss. This should reduce
repeated calls to glob().

Make use of bs_cache to simplify getting an unused index, for when we're
creating a backstore cfs entry. Remove _hbas() and _next_hba_index().

Signed-off-by: Andy Grover <agrover@redhat.com>