pscsi: Save udev_path instead of h:c:t:l
Both of hctl and /dev/sdX are bad for saving paths, since changes in
device enumeration ordering or udev may result in a name referring to a
different device. We should be using vpd83 or serial# or some other
immutable device id. Until we can do that, use the lesser of two evils;
removing/reinserting a drive may not even result in the same h:c:t:l, but
at least it will get the same sdX.

Signed-off-by: Andy Grover <agrover@redhat.com>