Add _configure() to StorageObject
Call it from derived objs' _configure method. This allows us to do common
post-enablement tasks in one place. One example is setting WWN, although
a no-op override method is needed for pscsi.

If we want to handle PR APTPL in StorageObject we also need this, since it
needs WWN set.

Signed-off-by: Andy Grover <agrover@redhat.com>