Add rtslib-saveconfig and rtslib-restoreconfig scripts
Targetcli is not scriptable. One way around this is to modify or create the
json describing the config directly, but even then, targetcli is needed to
save or restore the json file.

These new scripts enable saving or restoring a json-based configuration to
not rely on calling targetcli, or even being installed. These scripts and
just rtslib can properly return exit status & stderr output if
something goes wrong.

Signed-off-by: Andy Grover <agrover@redhat.com>