Add a 'clear' command to targetctl
It seems like this might be useful. If restore has errors, it will result
in a partially applied configuration. Maybe this what the user wants, or
maybe they'd just as soon have no configuration if any part of it fails.
If so, 'clear' to the rescue if exit status from 'restore' is nonzero.

Signed-off-by: Andy Grover <agrover@redhat.com>