Have rtslib mount configfs if not already mounted
The system service that restores target state may not always be configured
to run. If so, running targetcli may return strange error messages,
because it also is responsible for mounting configfs.

Instead, have rtslib try to mount configfs. If it's already mounted then
no problem, but this will ensure that using targetcli without the target
service activated will work, it just won't have the previous
configuration.

Signed-off-by: Andy Grover <agrover@redhat.com>