Add locking around configfs save/restore/clear operations
One concern raised on the list recently was that concurrent readers and
writers to configfs could result in things blowing up. This patch adds
a lock, which should serialize save, restore, and clear operations that
are used by targetctl, and potentially other clients via the rtslib API.

However, targetcli uses rtslib objects returned from other RTSRoot methods,
and could also potentially cause issues. We need to determine if we are OK
with that, or if more steps need to be taken to avoid this.

Signed-off-by: Andy Grover <agrover@redhat.com>