Don't call from_fabric_wwn from wwns()
from_fabric_wwn and wwns() do two slightly different things, and it's
better if we keep them separate. from/to_fabric_wwn() convert our
canonical wwn format from/to the format the fabric in configfs expects.
wwns() reads from somewhere else to get a list of wwns. The format of
WWNs may not be the same as what the fabric uses, and it should convert
whatever it reads to canonical format without referencing what the
fabric's WWN format is.

I really hope this is the last ever commit on this stupid cesspool of
my own making.

Signed-off-by: Andy Grover <agrover@redhat.com>