Do not check WWN is present in normalize_wwn
For tcm_fc and maybe others, it should be possible to configure a target
for a hw-based WWN before the WWN shows up in the list of WWNs. The kernel
code allows and handles this case.

The only downside is that a user might have a typo in entering a hw wwn by
hand, and experience confusion when it is allowed but things don't work
right. Clients may wish to warn the user if so.

Signed-off-by: Andy Grover <agrover@redhat.com>