Change set_parameters/attributes to take an err_func
We don't want to silently hide errors when setting things, but we also
don't want to stop. Since these are both called from setup() functions,
we can just pass an err_func to them and they can call it if there's a
problem, but keep going.

When calling set_attribute for the storageobject, use a custom err_func
that also includes the name and plugin in the message.

Signed-off-by: Andy Grover <agrover@redhat.com>