rtslib.Config can load unknown configfs attributes
Attributes not described in policy files but existing in the installed kernel
modules can now be loaded. They default to a 'raw' type, with a default value
equal to their value in configfs. This enables to use rtslib.Config on bleeding
edge kernels with new attribuites not described in policy yet.