Consolidate two mapping dicts into one
The storageobjects and "mapping" dict in so_from_path are almost the same,
except for name/path divergences due to alt_dirprefix. Consolidate these
into one list, which works with a few duplicate entries.

Put the new so_mapping dict next to bs_params dict, so the places to touch
for new backstores are a little more centralized.

Signed-off-by: Andy Grover <agrover@redhat.com>