Save/restore alias values for LUN and MappedLUN if present
We generate aliases -- the symlink names that link LUNs to storage
objects, and MappedLUNs to LUNs -- using uuid, and generating a uuid
needs the system to have a seeded source of randomness.

If we're restoring a config on boot, we may not have enough randomness and
this can cause long delays.

Instead, save alias values (by including them in dump()), and then restore
them if present. If not present, then a fresh alias value is generated.

See https://github.com/agrover/targetcli-fb/issues/58

Reported-by: Timofey <nefelim4ag@gmail.com>
Signed-off-by: Andy Grover <agrover@redhat.com>