Outline of an approach using pyudev.
This is an outline of what the approach would look like.

Some notes:
1. That messy "is this a partition" bit in get_size_for_disk_name might
be something that pyudev should learn to make itself responsible for.
I don't understand the details, so I'm not certain.
2. This version works against current Fedora package which is two minor versions
behind. Generally speaking, exceptions should be a lot less abundant and
more organized in current version.
3. The toint() method will have to be replaced with a special conversion method
in major release 1. This is due to a design problem in current version.
4. Note that there is far more complete error handling in pyudev, but taking
into account possible errors makes some methods more complicated.
5. Currently, pyudev can do nothing about the scsi/hctl functions that you
have, AFAIK. But, it's possible that it should be able to.