Make sure to properly avoid treating partitions the same as disk.
Return None if device path is a partition in convert_scsi_path_to_hctl.

In convert_scsi_hctl_to_path ensure that device is not a partition in the
most correct and efficient way. Get exactly the set of devices that are
descendants of the scsi device and are block devices and have type disk.
This should be a singleton, or the empty set.

Signed-off-by: mulhern <amulhern@redhat.com>