Rework convert_scsi_path_to_hctl and convert_scsi_hctl_to_path
Do not check for type=disk, this is not a requirement.

Raise an RTSLibError if not found, rather than returning None.

Rework PSCSIStorageObject._configure to use the new semantics, and avoid
else clauses in try blocks, since they are confusing and not strictly
needed.

Properly qualify DeviceNotFoundError as pyudev.DeviceNotFoundError.

Signed-off-by: Andy Grover <agrover@redhat.com>