ALUA: return list of members v2
There is a bug in the kernel where if you map a backend device through
multiple targets/tpgts it only returns the first one. The rtslib
ALUA members function then returns a string containing whatever the
kernel
prints out which is just the first lun the backend is mapped to. The
others are dropped. When the kernel is fixed this will print out a lun
mapping per line with the format:

driver/target/tpgt/lun

for each line.

This patch prepares us for the kernel getting fixed and returns
the lun mappings in a format that is easier for the rtslib user.

It will return a list of dictionarys with the format:

[{'tpgt': 1, 'driver': 'iSCSI', 'target':
'iqn.1999-09.com.tcmu:alua', 'lun': 3}, {'tpgt': 2,
'driver': 'iSCSI', 'target': 'iqn.1999-09.com.tcmu:alua2', 'lun': 1}]

--
v2 - remove tpgt/lun prefixes and return ints instead of strings.

Signed-off-by: Mike Christie <mchristi@redhat.com>