Switch target driver DB root dir to /etc/target
This switches the kernel target driver directory from
the default of /var/target to /etc/target, if the
"dbroot" sysfs attribute is present. If not, the default
of /var/target is maintained. This has to be done
by rtslib/root.py since this module loads the
target_core_mod module, where the dbroot value
must be updated before any target_core_* drivers
register with target_core_mod.

The switching is done via a new configuration file in
/etc/target called rtslib.cfg, which allows
this directory location to be modified.