Do not tie setting StorageObject wwn to enablement
This patch allows the wwn to be set before or after
the device has been enabled. It will be needed for
the delayed tcmu enablement feature where you would
normally create the device then set the attrs like
max_data_area_mb, and wwn then enable it.

Note:

I could not figure out why we required the device to be
enabled before setting the serial The kernel looks like it
does not care. The only kernel restriction looks like you
cannot have exported the device as a lun then set
the serial.

Signed-off-by: Mike Christie <mchristi@redhat.com>