Allow tcmu to delay enablement of its devices
tcmu seems to have a lot of funky settings that are needed
before enablement like preallocated buffers for the ring.
This patch allows us to create the object, but not yet
enable it.

It also allows us to create the device and retry the enablement
process which for tcmu-runner can commonly fail due to temp transport
issues like the network being down for the runner handler. In this
case during system startup we might want to create the device, retry
the enable step N times and then either destroy the object like we
currently do or leave it in the system so the admin can retry manually.
For the latter, we may not want to completely destroy the device
because the admin will not see it in targetcli and have to resetup
the backend and also the lun exports. With this patch they will see
all that in the targetcli but "enable" will show as false.

Note that for targetcli someone that cares about the second issue
should add some code so in targetcli we see the enabled value
in the default output so it is easier for the user like this:

  | o- user:file ...................................... [Storage Objects: 1]
  | | o- filetestsync ......... [/work/luns/lun0 (1.0GiB) activated enabled]

Looking at you gluster since I did the base work for you :)

Signed-off-by: Mike Christie <mchristi@redhat.com>