Allow bs_cache to be bypassed
The ceph tools might have multiple daemons making and querying the system.
To avoid invalidating the cache for every creation and lookup this allows
tools like ours to avoid the cache.

If the caller passed in the index during creations or lookups then use
that value instead of building/quering the cache.

Signed-off-by: Mike Christie <mchristi@redhat.com>