Handle attribute pi_prot_format being write-only
A recent kernel changes (see commit 6baca7601bdee2e5) make
the pi_prot_format protection information attribute write-only,
since reading it always returned 0.

But rtslib tries to read everything in the "attrib" (attribute)
directory, so reading pi_prot_format raises an access error,
since reading a write-only file is not allowed. So catch this
exception and handle it, assuming the value is 0 if we cannot
read the attribute.