Add 'readable' param to Group list_*() methods
Extend use of the new 'readable' optional param
for the Node class into the Group class, so they
also return writable *and* readable entries.

Did NOT update the calls to list_attributes() and
list_parameters() in NodeACLGroup class, since
we don't need ACLs to be readable to write them
out.