Filter fabric wwns by reading driver name
symbolic_name [1] is an attribute that represents any human-readable
string that means the name of a node or a port. This string can be
formatted in many different ways:

 - <HBA Model> FW: <Firmware Level> DVR: <Driver Level>
 - <HBA Model> FW: <Firmware Level> DVR: <Driver Level> - <Function>
 - <Manufacturer> <Device Type> <Model>
 - <Port Name>: <Vendor Name>: <Model> [<Device Number>]

Storage vendors can change symbolic_name for their needs to represent
array port name or array name. It is not a persistent way to determine
if a WWN is supported in fabric module. A more reliable way is to check
driver name.

1. https://standards.incits.org/apps/group_public/download.php/82426/T11-2016-046v0.pdf

Signed-off-by: Anastasia Kovaleva <a.kovaleva@yadro.com>