dbroot: let the user change the dbroot directory freely.
Thanks to a new kernel patch, the LIO driver now lets the user change
the dbroot directory even if some target modules are already registered
(if some conditions are met).

rtslib should let the user set their preferred root directory without
arbitrary checks and let the kernel decide if this is allowed or not.

Signed-off-by: Maurizio Lombardi <mlombard@redhat.com>